start on js SwapChain
